[feat] add koin librarie for depedency injection
