+++ RickAndMortyApp/app/src/main/java/gonzalez/oscar/rickandmortyapp/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package gonzalez.oscar.rickandmortyapp
+package gonzalez.oscar.rickandmortyapp.presentation
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
@@ -6,18 +6,20 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
+import gonzalez.oscar.rickandmortyapp.R.id
+import gonzalez.oscar.rickandmortyapp.R.layout
 import kotlinx.android.synthetic.main.activity_main.nav_view
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(layout.activity_main)
 
-        val navController = findNavController(R.id.nav_host_fragment)
+        val navController = findNavController(id.nav_host_fragment)
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_characters, R.id.navigation_episodes, R.id.navigation_locations
+                id.navigation_characters, id.navigation_episodes, id.navigation_locations
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
