+++ RickAndMortyApp/app/src/main/java/gonzalez/oscar/rickandmortyapp/presentation/MainApplication.kt
@@ -1,0 +1,17 @@
+package gonzalez.oscar.rickandmortyapp.presentation
+
+import android.app.Application
+import gonzalez.oscar.rickandmortyapp.di.modulesRickAndMorty
+import org.koin.android.ext.koin.androidContext
+import org.koin.core.context.startKoin
+
+class MainApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        startKoin {
+            androidContext(this@MainApplication)
+            modules(modulesRickAndMorty)
+        }
+    }
+}
