+++ RickAndMortyApp/app/src/main/java/gonzalez/oscar/rickandmortyapp/presentation/ui/characters/CharactersViewModel.kt
@@ -1,13 +1,13 @@
-package gonzalez.oscar.rickandmortyapp.ui.notifications
+package gonzalez.oscar.rickandmortyapp.presentation.ui.characters
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class LocationsViewModel : ViewModel() {
+class CharactersViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is locations Fragment"
+        value = "This is characters Fragment"
     }
     val text: LiveData<String> = _text
 }
