+++ RickAndMortyApp/app/src/main/java/gonzalez/oscar/rickandmortyapp/presentation/ui/episodes/EpisodesFragment.kt
@@ -1,4 +1,4 @@
-package gonzalez.oscar.rickandmortyapp.ui.dashboard
+package gonzalez.oscar.rickandmortyapp.presentation.ui.episodes
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,20 +7,18 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import gonzalez.oscar.rickandmortyapp.R
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class EpisodesFragment : Fragment() {
 
-    private lateinit var mEpisodesViewModel: EpisodesViewModel
+    private val mEpisodesViewModel: EpisodesViewModel by viewModel()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        mEpisodesViewModel =
-            ViewModelProvider(this).get(EpisodesViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_episodes, container, false)
         val textView: TextView = root.findViewById(R.id.text_dashboard)
         mEpisodesViewModel.text.observe(viewLifecycleOwner, Observer {
