+++ RickAndMortyApp/app/src/main/java/gonzalez/oscar/rickandmortyapp/presentation/ui/locations/LocationsFragment.kt
@@ -1,4 +1,4 @@
-package gonzalez.oscar.rickandmortyapp.ui.notifications
+package gonzalez.oscar.rickandmortyapp.presentation.ui.locations
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,20 +7,18 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import gonzalez.oscar.rickandmortyapp.R
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class LocationsFragment : Fragment() {
 
-    private lateinit var mLocationsViewModel: LocationsViewModel
+    private val mLocationsViewModel: LocationsViewModel by viewModel()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        mLocationsViewModel =
-            ViewModelProvider(this).get(LocationsViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_locations, container, false)
         val textView: TextView = root.findViewById(R.id.text_notifications)
         mLocationsViewModel.text.observe(viewLifecycleOwner, Observer {
