+++ RickAndMortyApp/app/src/main/java/gonzalez/oscar/rickandmortyapp/presentation/ui/locations/LocationsViewModel.kt
@@ -1,13 +1,13 @@
-package gonzalez.oscar.rickandmortyapp.ui.dashboard
+package gonzalez.oscar.rickandmortyapp.presentation.ui.locations
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class EpisodesViewModel : ViewModel() {
+class LocationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is episodes Fragment"
+        value = "This is locations Fragment"
     }
     val text: LiveData<String> = _text
 }
