+++ RickAndMortyApp/app/src/main/res/layout/fragment_characters.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.CharactersFragment">
+    tools:context=".presentation.ui.characters.CharactersFragment">
 
     <TextView
         android:id="@+id/text_home"
