+++ RickAndMortyApp/app/src/main/res/layout/fragment_locations.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notifications.LocationsFragment">
+    tools:context=".presentation.ui.locations.LocationsFragment">
 
     <TextView
         android:id="@+id/text_notifications"
